requirement.txt
aniso8601==8.0.0
click==7.1.1
crcmod==1.7
Flask==1.1.2
getmac==0.8.2
intelhex==2.2.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
pymysensors==0.20.1
pyserial==3.4
pyserial-asyncio==0.4
pytz==2019.3
six==1.14.0
voluptuous==0.11.7
Werkzeug==1.0.1
